IGNITE-6858 wip.
